Improve ref generation.

+++ pipelines/cnv/src/step1-masking/run-pipe.sh
@@ -1,39 +1,78 @@
 #!/bin/bash
-# vim:set ts=2 sw=2 et paste:
+# vim:set ts=2 sw=2 et paste foldmethod=indent:
 #
 # In case you want to remove:
 # rm -rf anno_kmer_masked.fa *.nohit  .RData  kmerCountsWithMrsFast_more20placements.txt  moab_logs/ deps/  *.bed chr*.fa counts.chr* *.sam all*.counts
 #
-[ ! -f "./config.sh" ] && echo "I can't find ./config.sh" && exit 1
-src_dir=$(dirname $(readlink -f $0))
-source ./config.sh
-[ ! -f $fasta ] && echo "I can't find $fasta" && exit 1
-[ ! -f $fasta_anno_masked ] && echo "I can't find $fasta" && exit 1
-
+# We need:
+: << EOF
+  wget http://hgdownload.cse.ucsc.edu/goldenPath/rheMac2/bigZips/chromFa.tar.gz
+  tar zxf chromFa.tar.gz
+  cat *.fa > raw_fasta.fa
+EOF
 
 pipe="bash" # "" to not submit or "bash" to submit to cluster
 th_over="20"
+src_dir=$(dirname $(readlink -f $0))
 
-        
-chrm_info="chrm_info.bed"
+
+error() {
+  local msg=$1
+  [ ".$msg" != "." ] && echo "ERROR: $msg"
+  exit 1
+}
+
+
+usage() {
+  local msg=$1
+  [ ".$msg" != "." ] && echo  "ERROR: $msg"
+  echo "Usage: $(basename $0) <raw_fasta.fa> <ucsc_genome>"
+  [ ".$msg" != "." ] && exit 1
+}
+
+
+# Params checks
+raw_fasta=$1
+ucsc_genome=$2
+[ ".$raw_fasta" == "." ] && usage "I need raw fasta name"
+[ ".$ucsc_genome" == "." ] && usage "Need ucsc genome code"
+
+
+# default filename outputs. Don't change.
 fasta_anno_masked="fasta_anno_masked.fa"
 fa_without_over="anno_kmer_masked.fa"
+chrm_info="chrm_info.bed"
 
-mysql --user=genome --host=genome-mysql.cse.ucsc.edu -A \
-    -e "select chrom, size from $ucsc_genome.chromInfo" | \
-    grep chr | awk '{print $1"\t"$2"\t"$2}' | grep -v size > $chrm_info
 
-[ ! -f $chrm_info ] && echo "I can't find $chrm_info" && exit 1
+# STEP1: download chrm info for genome
+########
+if [ ! -f $chrm_info.bed ];then
+  mysql --user=genome --host=genome-mysql.cse.ucsc.edu -A \
+      -e "select chrom, size from ${ucsc_genome}.chromInfo" | \
+      grep chr | awk '{print $1"\t"$2"\t"$2}' | grep -v size > $chrm_info
+else
+  echo "$chrm_info already there."
+fi
+
+[ ! -f $chrm_info ] && error "I can't find $chrm_info" && exit 1
 rm -rf deps; mkdir deps
+for i in `seq 1 10`;do
+  touch deps/${i}.txt
+done
 
 
-# You may want to run this manually.
-# Or, better, we should make it more generic
+# STEP2: Download anotations and apply them
 ###################################
-cmd="$src_dir/mask_rtf_simple_gaps.sh $fasta $fasta_anno_masked"
-submit -e -s "gen_masked_ref" $cmd | $pipe >> deps/1.txt
+if [ ! -f ./$fasta_anno_masked ];then
+  cmd="$src_dir/mask_rtf_simple_gaps.sh $raw_fasta $fasta_anno_masked $ucsc_genome"
+  submit -e -s "gen_masked_ref" $cmd | $pipe >> deps/1.txt
+else
+  echo "$fasta_anno_masked already there."
+fi
 
 
+# STEP3: Generate read coordinates in genome
+######################################
 while read line;do
     chrm=$(echo -e $line | cut -f1 -d' ')
     len=$(echo -e $line | cut -f2 -d' ')
@@ -42,6 +81,8 @@ while read line;do
 done < $chrm_info
 
 
+# STEP4: extract reads from genome
+######################################
 while read line;do
     chrm=$(echo -e $line | cut -f1 -d' ')
     len=$(echo -e $line | cut -f2 -d' ')
@@ -52,6 +93,8 @@ while read line;do
 done < $chrm_info
 
 
+# STEP5: map genome reads to genome
+####################################
 while read line;do
     chrm=$(echo -e $line | cut -f1 -d' ')
     reads="${chrm}.reads.fa"
@@ -59,11 +102,14 @@ while read line;do
     # FIXME
     # Hack to avoid failing dependency.
     # There is one read file with all N's and mrsfast exit != 0
-    cmd="mrsfast --search $fasta_anno_masked --seq $reads -o $out -e 2; exit 0"
+    mrsfast="$src_dir/../../bin/mrsfast"
+    cmd="$mrsfast --search $fasta_anno_masked --seq $reads -o $out -e 2; exit 0"
     cat deps/3.txt | submit -e -f - -m 16G -s "map_kmers.$chrm" "$cmd" | $pipe >> deps/4.txt
 done < $chrm_info
 
 
+# STEP6: Count how many alignments we have per genome location
+####################################
 while read line;do
     chrm=$(echo -e $line | cut -f1 -d' ')
     kmer_file="${chrm}.reads.fa"
@@ -77,9 +123,12 @@ while read line;do
 done < $chrm_info
 
 
+# STEP7: merge counts
+########################
 cmd="$src_dir/process_counts.sh $th_over"
 cat deps/5.txt | submit -e -f - -s "working_on_counts" $cmd | $pipe >> deps/6.txt
 
-
+# STEP8: Mask overrepresented regions in the annotated fasta reference
+##############################
 cmd="maskFastaFromBed -fi $fasta_anno_masked -bed all.$th_over.counts -fo $fa_without_over"
 cat deps/6.txt | submit -e -f - -s "mask_final" "$cmd" | $pipe
